requirements.minimal.txt
gradio==5.33.1
langgraph==0.4.8
langchain==0.3.25
langchain-openai
python-dotenv
playwright
requests
pydantic
langchain-community
